Added so much stuff.
Fixed the bug where the heal room sometimes won't spawn. Now it will spawn with a higher % chance. But the chance of it getting a 0% chance of spawning is still there. Gotta fix that in the next update. Random level generation is all good. Gotta fix the heal and exit room.